SKOTE-1: update login flow
